#6 move validator to middleware
